+++ dev/src/v1/firestore_client.ts
@@ -343,7 +343,7 @@ export class FirestoreClient {
   // -- Service calls --
   // -------------------
   getDocument(
-    request: protos.google.firestore.v1.IGetDocumentRequest,
+    request?: protos.google.firestore.v1.IGetDocumentRequest,
     options?: CallOptions
   ): Promise<
     [
@@ -398,7 +398,7 @@ export class FirestoreClient {
    * const [response] = await client.getDocument(request);
    */
   getDocument(
-    request: protos.google.firestore.v1.IGetDocumentRequest,
+    request?: protos.google.firestore.v1.IGetDocumentRequest,
     optionsOrCallback?:
       | CallOptions
       | Callback<
@@ -437,7 +437,7 @@ export class FirestoreClient {
     return this.innerApiCalls.getDocument(request, options, callback);
   }
   updateDocument(
-    request: protos.google.firestore.v1.IUpdateDocumentRequest,
+    request?: protos.google.firestore.v1.IUpdateDocumentRequest,
     options?: CallOptions
   ): Promise<
     [
@@ -498,7 +498,7 @@ export class FirestoreClient {
    * const [response] = await client.updateDocument(request);
    */
   updateDocument(
-    request: protos.google.firestore.v1.IUpdateDocumentRequest,
+    request?: protos.google.firestore.v1.IUpdateDocumentRequest,
     optionsOrCallback?:
       | CallOptions
       | Callback<
@@ -537,7 +537,7 @@ export class FirestoreClient {
     return this.innerApiCalls.updateDocument(request, options, callback);
   }
   deleteDocument(
-    request: protos.google.firestore.v1.IDeleteDocumentRequest,
+    request?: protos.google.firestore.v1.IDeleteDocumentRequest,
     options?: CallOptions
   ): Promise<
     [
@@ -585,7 +585,7 @@ export class FirestoreClient {
    * const [response] = await client.deleteDocument(request);
    */
   deleteDocument(
-    request: protos.google.firestore.v1.IDeleteDocumentRequest,
+    request?: protos.google.firestore.v1.IDeleteDocumentRequest,
     optionsOrCallback?:
       | CallOptions
       | Callback<
@@ -624,7 +624,7 @@ export class FirestoreClient {
     return this.innerApiCalls.deleteDocument(request, options, callback);
   }
   beginTransaction(
-    request: protos.google.firestore.v1.IBeginTransactionRequest,
+    request?: protos.google.firestore.v1.IBeginTransactionRequest,
     options?: CallOptions
   ): Promise<
     [
@@ -672,7 +672,7 @@ export class FirestoreClient {
    * const [response] = await client.beginTransaction(request);
    */
   beginTransaction(
-    request: protos.google.firestore.v1.IBeginTransactionRequest,
+    request?: protos.google.firestore.v1.IBeginTransactionRequest,
     optionsOrCallback?:
       | CallOptions
       | Callback<
@@ -713,7 +713,7 @@ export class FirestoreClient {
     return this.innerApiCalls.beginTransaction(request, options, callback);
   }
   commit(
-    request: protos.google.firestore.v1.ICommitRequest,
+    request?: protos.google.firestore.v1.ICommitRequest,
     options?: CallOptions
   ): Promise<
     [
@@ -764,7 +764,7 @@ export class FirestoreClient {
    * const [response] = await client.commit(request);
    */
   commit(
-    request: protos.google.firestore.v1.ICommitRequest,
+    request?: protos.google.firestore.v1.ICommitRequest,
     optionsOrCallback?:
       | CallOptions
       | Callback<
@@ -803,7 +803,7 @@ export class FirestoreClient {
     return this.innerApiCalls.commit(request, options, callback);
   }
   rollback(
-    request: protos.google.firestore.v1.IRollbackRequest,
+    request?: protos.google.firestore.v1.IRollbackRequest,
     options?: CallOptions
   ): Promise<
     [
@@ -850,7 +850,7 @@ export class FirestoreClient {
    * const [response] = await client.rollback(request);
    */
   rollback(
-    request: protos.google.firestore.v1.IRollbackRequest,
+    request?: protos.google.firestore.v1.IRollbackRequest,
     optionsOrCallback?:
       | CallOptions
       | Callback<
@@ -889,7 +889,7 @@ export class FirestoreClient {
     return this.innerApiCalls.rollback(request, options, callback);
   }
   batchWrite(
-    request: protos.google.firestore.v1.IBatchWriteRequest,
+    request?: protos.google.firestore.v1.IBatchWriteRequest,
     options?: CallOptions
   ): Promise<
     [
@@ -950,7 +950,7 @@ export class FirestoreClient {
    * const [response] = await client.batchWrite(request);
    */
   batchWrite(
-    request: protos.google.firestore.v1.IBatchWriteRequest,
+    request?: protos.google.firestore.v1.IBatchWriteRequest,
     optionsOrCallback?:
       | CallOptions
       | Callback<
@@ -989,7 +989,7 @@ export class FirestoreClient {
     return this.innerApiCalls.batchWrite(request, options, callback);
   }
   createDocument(
-    request: protos.google.firestore.v1.ICreateDocumentRequest,
+    request?: protos.google.firestore.v1.ICreateDocumentRequest,
     options?: CallOptions
   ): Promise<
     [
@@ -1048,7 +1048,7 @@ export class FirestoreClient {
    * const [response] = await client.createDocument(request);
    */
   createDocument(
-    request: protos.google.firestore.v1.ICreateDocumentRequest,
+    request?: protos.google.firestore.v1.ICreateDocumentRequest,
     optionsOrCallback?:
       | CallOptions
       | Callback<
@@ -1247,7 +1247,7 @@ export class FirestoreClient {
   }
 
   listDocuments(
-    request: protos.google.firestore.v1.IListDocumentsRequest,
+    request?: protos.google.firestore.v1.IListDocumentsRequest,
     options?: CallOptions
   ): Promise<
     [
@@ -1326,7 +1326,7 @@ export class FirestoreClient {
    *   for more details and examples.
    */
   listDocuments(
-    request: protos.google.firestore.v1.IListDocumentsRequest,
+    request?: protos.google.firestore.v1.IListDocumentsRequest,
     optionsOrCallback?:
       | CallOptions
       | PaginationCallback<
@@ -1514,7 +1514,7 @@ export class FirestoreClient {
     ) as AsyncIterable<protos.google.firestore.v1.IDocument>;
   }
   partitionQuery(
-    request: protos.google.firestore.v1.IPartitionQueryRequest,
+    request?: protos.google.firestore.v1.IPartitionQueryRequest,
     options?: CallOptions
   ): Promise<
     [
@@ -1602,7 +1602,7 @@ export class FirestoreClient {
    *   for more details and examples.
    */
   partitionQuery(
-    request: protos.google.firestore.v1.IPartitionQueryRequest,
+    request?: protos.google.firestore.v1.IPartitionQueryRequest,
     optionsOrCallback?:
       | CallOptions
       | PaginationCallback<
@@ -1804,7 +1804,7 @@ export class FirestoreClient {
     ) as AsyncIterable<protos.google.firestore.v1.ICursor>;
   }
   listCollectionIds(
-    request: protos.google.firestore.v1.IListCollectionIdsRequest,
+    request?: protos.google.firestore.v1.IListCollectionIdsRequest,
     options?: CallOptions
   ): Promise<
     [
@@ -1859,7 +1859,7 @@ export class FirestoreClient {
    *   for more details and examples.
    */
   listCollectionIds(
-    request: protos.google.firestore.v1.IListCollectionIdsRequest,
+    request?: protos.google.firestore.v1.IListCollectionIdsRequest,
     optionsOrCallback?:
       | CallOptions
       | PaginationCallback<