+++ types/firestore.d.ts
@@ -28,9 +28,17 @@ declare namespace FirebaseFirestore {
   export type DocumentData = {[field: string]: any};
 
   /**
-   * Update data (for use with `DocumentReference.update()`) consists of field
-   * paths (e.g. 'foo' or 'foo.baz') mapped to values. Fields that contain dots
-   * reference nested fields within the document.
+   * Update data (for use with [update]{@link DocumentReference#update})
+   * that contains paths mapped to values. Fields that contain dots reference
+   * nested fields within the document.
+   *
+   * You can update a top-level field in your document by using the field name
+   * as a key (e.g. `foo`). The provided value completely replaces the contents
+   * for this field.
+   *
+   * You can also update a nested field directly by using its field path as a
+   * key (e.g. `foo.bar`). This nested field update replaces the contents at
+   * `bar` but does not modify other data under `foo`.
    */
   export type UpdateData = {[fieldPath: string]: any};
 
@@ -164,6 +172,28 @@ declare namespace FirebaseFirestore {
     ignoreUndefinedProperties?: boolean;
 
     [key: string]: any; // Accept other properties, such as GRPC settings.
+  }
+
+  /** Options to configure a read-only transaction. */
+  export interface ReadOnlyTransactionOptions {
+    /** Set to true to indicate a read-only transaction. */
+    readOnly: true;
+    /**
+     * If specified, documents are read at the given time. This may not be more
+     * than 60 seconds in the past from when the request is processed by the
+     * server.
+     */
+    readTime?: Timestamp;
+  }
+
+  /** Options to configure a read-write transaction. */
+  export interface ReadWriteTransactionOptions {
+    /** Set to false or omit to indicate a read-write transaction. */
+    readOnly?: false;
+    /**
+     * The maximum number of attempts for this transaction. Defaults to five.
+     */
+    maxAttempts?: number;
   }
 
   /**
@@ -300,14 +330,33 @@ declare namespace FirebaseFirestore {
      * the transaction.
      *
      * You can use the transaction object passed to 'updateFunction' to read and
-     * modify Firestore documents under lock. Transactions are committed once
-     * 'updateFunction' resolves and attempted up to five times on failure.
+     * modify Firestore documents under lock. You have to perform all reads
+     * before before you perform any write.
+     *
+     * Transactions can be performed as read-only or read-write transactions. By
+     * default, transactions are executed in read-write mode.
+     *
+     * A read-write transaction obtains a pessimistic lock on all documents that
+     * are read during the transaction. These locks block other transactions,
+     * batched writes, and other non-transactional writes from changing that
+     * document. Any writes in a read-write transactions are committed once
+     * 'updateFunction' resolves, which also releases all locks.
+     *
+     * If a read-write transaction fails with contention, the transaction is
+     * retried up to five times. The `updateFunction` is invoked once for each
+     * attempt.
+     *
+     * Read-only transactions do not lock documents. They can be used to read
+     * documents at a consistent snapshot in time, which may be up to 60 seconds
+     * in the past. Read-only transactions are not retried.
+     *
+     * Transactions time out after 60 seconds if no documents are read.
+     * Transactions that are not committed within than 270 seconds are also
+     * aborted. Any remaining locks are released when a transaction times out.
      *
      * @param updateFunction The function to execute within the transaction
      * context.
-     * @param {object=} transactionOptions Transaction options.
-     * @param {number=} transactionOptions.maxAttempts The maximum number of
-     * attempts for this transaction.
+     * @param transactionOptions Transaction options.
      * @return If the transaction completed successfully or was explicitly
      * aborted (by the updateFunction returning a failed Promise), the Promise
      * returned by the updateFunction will be returned here. Else if the
@@ -316,7 +365,9 @@ declare namespace FirebaseFirestore {
      */
     runTransaction<T>(
       updateFunction: (transaction: Transaction) => Promise<T>,
-      transactionOptions?: {maxAttempts?: number}
+      transactionOptions?:
+        | ReadWriteTransactionOptions
+        | ReadOnlyTransactionOptions
     ): Promise<T>;
 
     /**